pyproject.toml
[tool.poetry]
name = "Running_RESTful_API"
version = "0.1.0"
description = ""
authors = ["Clark Rutter <c.rutter50@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-restful = "^0.3.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pytest = "^8.2.2"
pytest-flask = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
